requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
Django==3.2.14
django-ckeditor==6.4.2
django-cors-headers==3.13.0
django-filter==22.1
django-js-asset==2.0.0
django-polymorphic==3.1.0
django-versatileimagefield==2.2
djangorestframework==3.13.1
generics==3.8.0
idna==3.3
Markdown==3.4.1
Pillow==9.2.0
psycopg2==2.9.3
python-magic==0.4.27
python-magic-bin==0.4.14
pytz==2022.1
requests==2.28.1
slacker==0.14.0
sqlparse==0.4.2
urllib3==1.26.10
